# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vehicle-Service API",
        "description": "REST API untuk Service Data Kendaraan pada proses Penugasan Perjalanan Dinas (Dispatching).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Vehicle-Service local server"
        }
    ],
    "paths": {
...
        "/api/v1/vehicles/{id}/dispatch": {
            "post": {
                "tags": [
                    "Tugas 3 - Dispatch"
                ],
                "summary": "Transaksi kritis penugasan kendaraan dengan SSO, SOAP audit, dan RabbitMQ event",
                "operationId": "17baf04db7295c29fb20241cd4912fd5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "trip_reference",
                                    "requester_name",
                                    "destination",
                                    "start_date"
                                ],
                                "properties": {
                                    "trip_reference": {
                                        "type": "string",
                                        "example": "TRIP-T3-001"
                                    },
                                    "requester_name": {
                                        "type": "string",
                                        "example": "Admin Perjalanan Dinas"
                                    },
                                    "destination": {
                                        "type": "string",
                                        "example": "Kantor Cabang Bandung"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-10"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-11",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dispatch berhasil, SOAP receipt tersimpan, event RabbitMQ dipublish"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Bearer JWT kosong atau tidak valid"
                    },
                    "403": {
                        "description": "Role SSO tidak diizinkan"
                    },
                    "404": {
                        "description": "Kendaraan tidak ditemukan"
                    },
                    "409": {
                        "description": "Kendaraan tidak Available"
                    }
                },
                "security": [
                    {
                        "ssoBearer": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "iaeApiKey": {
                "type": "apiKey",
                "description": "API Key untuk mengakses Vehicle-Service. Gunakan value 102022400230.",
                "name": "X-IAE-KEY",
                "in": "header"
            },
            "ssoBearer": {
                "type": "http",
                "description": "Bearer JWT dari Cloud SSO Dosen.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}